moved search and profile components from app component directory to their respective component directories and updated paths accordingly
