+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/AthleteProfile.tsx
