+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/AthleteProfile_leg.tsx
@@ -5,9 +5,9 @@
 import { useEffect, useState } from "react";
 import { SafeAthlete } from "@/app/types";
 import Container from "@/app/components/Container";
-import AthleteHead from "@/app/components/profile/ProfileHead";
-import AthleteInfo from "@/app/components/profile/ProfileInfo";
-import Description from "@/app/components/profile/Description";
+import AthleteHead from "@/app/(dashboard)/(routes)/profiles/[type]/[id]/components/ProfileHead";
+import AthleteInfo from "@/app/(dashboard)/(routes)/profiles/[type]/[id]/components/ProfileInfo";
+import Description from "@/app/(dashboard)/(routes)/profiles/[type]/[id]/components/Description";
 import CollaborationRequest from "./CollaborationRequest";
 
 interface AthleteProfileProps {
