+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/BrandProfile.tsx
