+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/PredicateSearch.tsx
@@ -4,7 +4,7 @@
 
 import React, { useState, useEffect, useMemo } from 'react';
 import TripleCardPredicate from './TripleCardPredicate';
-import Loader from '../Loader';
+import Loader from '../../../../../../components/Loader';
 import Masonry from 'react-masonry-css'; // Use react-masonry-css for React compatibility
 import debounce from 'lodash.debounce';
 import { useRouter, useSearchParams } from 'next/navigation';
