+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/PredicateSidebar.tsx
