+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/Profile.css
