+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/ProfileGrid.tsx
