+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/ProfileHead.tsx
