+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/ProfileInfo.tsx
