+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/SkyboxBanner.tsx
