+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/TripleCardObject.tsx
