+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/TripleCardPredicate.tsx
