+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/TripleCardSubject.tsx
