+++ app/(dashboard)/(routes)/profiles/[type]/[id]/components/TripleCardWhole.tsx
