+++ app/(dashboard)/(routes)/search/components/AIModelSearchCard.tsx
