+++ app/(dashboard)/(routes)/search/components/ActiveFiltersSection.tsx
