+++ app/(dashboard)/(routes)/search/components/AvatarAndUsername.tsx
