+++ app/(dashboard)/(routes)/search/components/BrandSearchCard.tsx
