+++ app/(dashboard)/(routes)/search/components/ContractCardTermDetail.tsx
