+++ app/(dashboard)/(routes)/search/components/CustomSearchBox.tsx
