+++ app/(dashboard)/(routes)/search/components/ModelTypeIcon.tsx
