+++ app/(dashboard)/(routes)/search/components/SearchCardBadgeDetail.tsx
