+++ app/(dashboard)/(routes)/search/components/SearchPageLayout.tsx
