+++ app/(dashboard)/(routes)/search/components/SearchSimilarityComponent.tsx
