+++ app/(dashboard)/(routes)/search/components/SidebarToggle.tsx
